arm: Fix for highmem detection
Found on nvidia github. Only applied it via web interface



sanity_check_meminfo walks over the registered memory banks and attempts

to split banks across lowmem and highmem when they would otherwise

overlap with the vmalloc space.



The checks to determine whether or not a bank belongs to highmem

or not only check if __va(bank->start) is greater or less than

vmalloc_min. In the case that it is equal, the bank is incorrectly

treated as lowmem, which hoses the vmalloc area.



This patch fixes the issue by checking whether the virtual

start address of a bank is >= vmalloc_min



Similar fix is there in Linus' kernel tree for 3.0.0-rc2.

commit 40f7bfe4f1c2761abeceb3b2b9dc1feec3c47ed9



bug 807934



(cherry picked from commit 316bd372c8aa7513ea6dcc9b3f6ecb487de96a2a)



Change-Id: I1ffdaa5fb4c7c44c34d4a238e7e5e790eea572fc

Reviewed-on: http://git-master/r/34700

Tested-by: Manoj Chourasia <mchourasia@nvidia.com>

Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>